Create a class defining the Ticket Model
The Ticket class will be linked to the Project model with a many-to-one
relationship via ForeignKey.

This class will consist of a Title CharField section limited to 150
max characters, a Details text field, a Priority ChoiceField, a
DateTimeField that will automatically add the Date and Time from
creation, and a file upload FileFiled where users may add any optional
attachments, with support for multiple files at once.