Create form allowing for new projects to be made
Added in a forms model for users to create projects. Users will be able
to fill out the information for the project title and the details of the
project. The date and time will automatically be saved when the project
is submitted.

For details, a Textarea widget was used with the 'cols' attritbute set
to 80 to allow more room for the project to be outlined.