User authentication overhaul
Created a template users see upon successfully logging out.
Restricted Bugs views path to require users to login to view. Users who
are not logged in will be redirected to the log in page.
Migrated user data so that projects, tickets and comments are all now
associated to a user.