Added custom user model
Created a custom user model that defaults username to an email address.
Form will also include fields to fill a user profile with first and last
name, as well as track the date they were added.
Required flushing of database and new migrations to be created.
AUTH_USER_MODEL in settings.py has been updated to reflect this change