Implement Team model
Team model is connected to User class via foreign key relationship, so
many users can be connected to a single Team. Team model consists of
name, a manager user assigned upon creation, a description and the
creation date.

'date_joined' from User model needed to be assigned as a
readonly_field in order to be viewed in django's admin dashboard.

Logic for new project, new ticket and new comment bugs.views function
was reworked so that an owner was assigned upon creation. This
fixed an error when making a migration for the Team model addition.