Refactor User and CustomUserAdmin models
The User model has been changed in several areas:
1. The role field has been fixed to be an IntegerField, as needed
to be used with ROLE_CHOICES.
2. The team field has been changed to be nullable on creation.
Administrator users will be able to assign a team after creation
through the admin site.
3. The basic user role has been renamed to be the designer
role now.

UserManager now takes in the role field as a required variable.

CustomUserAdmin now has a more thorough custom fieldset
definition. Personal info includes email, first and last names,
and the date joined. Default info includes django default
variables is superuser, is staff, is active and groups. Roles
and permissions displays the user role, team, and the
role permission flags is administrator, is project manager,
is developer and is designer.

Finally, upon user creation the role permission flags are applied.
Active status is set to False to allow for implementation of an
email verification system to change to True after.