Add Teams, Team, and Team creation pages
Templates for a Teams directory page, a Team info page and a
Team creation page have been created. A path has been made
for the Teams page with a view function that is restricted to
users of administrator or project manager status.

A check_owner function has been added to further restrict users
from editing projects, tickets and comments which are not their
own.