Created a new endpoint for users profile view/edit
Users can now click on their name to be directed to a profile page where
they can view their current information and update it.
Role and Assigned teams cannot currently be modified, but support for
an Administrator or the team members manager to modify those feels will
later be added.