Refactor user privilege check for teams route
If the user is not an admin or manager, instead of seeing a list of teams
they will be shown their own teams page if they are assigned to one. If
not a 404 error will be raised.
Only admin's and managers will be shown a link to edit a team.